# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Swagger Documentation of SafeWorks server API",
    "version": "1.0.0",
    "title": "SafeWorks API"
  },
  "host": "localhost:3001",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/portfolios": {
      "post": {
        "tags": [
          "Portfolio"
        ],
        "description": "Endpoint to add a new Portfolio",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "portfolio",
            "in": "body",
            "description": "Porfolio data to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddPortfolio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created Portfolio object",
            "schema": {
              "$ref": "#/definitions/CreatedPorfolio"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Portfolio"
        ],
        "description": "Endpoint to get a Certification from Portfolio",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "Portfolio UID",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Certification ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/portfolios/{uid}/certifications": {
      "post": {
        "tags": [
          "Portfolio"
        ],
        "description": "Endpoint to add a new Certification on a Portfolio",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Portfolio UID"
          },
          {
            "name": "certification",
            "in": "body",
            "description": "Certification data to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddCertification"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Portfolio object with new Certification",
            "schema": {
              "$ref": "#/definitions/Porfolio"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/portfolios/{uid}": {
      "put": {
        "tags": [
          "Portfolio"
        ],
        "description": "Endpoint to update a Portfolio",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Portfolio UID"
          },
          {
            "name": "updatedPortfolio",
            "in": "body",
            "description": "Portfolio data to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePortfolio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated Portfolio object",
            "schema": {
              "$ref": "#/definitions/Porfolio"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Portfolio"
        ],
        "description": "Endpoint to get a Portfolio by UID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Portfolio UID"
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio object",
            "schema": {
              "$ref": "#/definitions/Porfolio"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Porfolio": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "portfolio": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "huHwBD3gTccU7pgA3Z94"
            },
            "user_uid": {
              "type": "string",
              "example": "NESmmmshXaSSeYHMB5FtU0l2hTg2"
            },
            "description": {
              "type": "string",
              "example": "Descrição"
            },
            "certifications": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "issue_date": {
                    "type": "string",
                    "example": "07/07/2023 18:32:28"
                  },
                  "description": {
                    "type": "string",
                    "example": "Descrição"
                  },
                  "id": {
                    "type": "string",
                    "example": "huHwBD3gTccU7pgA3Z941"
                  },
                  "title": {
                    "type": "string",
                    "example": "Certificado"
                  },
                  "issue_organization": {
                    "type": "string",
                    "example": "PUCPR"
                  },
                  "certification_url": {}
                }
              }
            },
            "years_experience": {
              "type": "number",
              "example": 1
            },
            "created": {
              "type": "string",
              "example": "07/07/2023 18:32:28"
            },
            "modified": {
              "type": "string",
              "example": "08/07/2023 18:32:28"
            },
            "deleted": {}
          }
        }
      }
    },
...
    "CreatedPorfolio": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "portfolio": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "huHwBD3gTccU7pgA3Z94"
            },
            "user_uid": {
              "type": "string",
              "example": "NESmmmshXaSSeYHMB5FtU0l2hTg2"
            },
            "description": {
              "type": "string",
              "example": "Descrição"
            },
            "certifications": {
              "type": "array",
              "example": [],
              "items": {}
            },
            "years_experience": {
              "type": "number",
              "example": 1
            },
            "created": {
              "type": "string",
              "example": "07/07/2023 18:32:28"
            },
            "modified": {
              "type": "string",
              "example": "08/07/2023 18:32:28"
            },
            "deleted": {}
          }
        }
      }
    },
    "AddCertification": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Certificado"
        },
        "description": {
          "type": "string",
          "example": "Descrição Certificado"
        },
        "issue_organization": {
          "type": "string",
          "example": "PUCR"
        },
        "issue_date": {
          "type": "string",
          "example": "2023-07-20"
        }
      },
      "required": [
        "title",
        "description",
        "issue_organization",
        "issue_date"
      ]
    },
    "UpdatePortfolio": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "Descrição"
        },
        "years_experience": {
          "type": "number",
          "example": 1
        }
      }
    },
...
  }
}